requirements.txt
asyncio==3.4.3
python-dotenv==1.0.0
google-generativeai==0.3.2
google-adk==0.1.0
logging==0.5.1.2 